+++ gridsome/__tests__/__fixtures__/project-path-prefix/gridsome.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   siteName: 'Gridsome',
-  pathPrefix: '/subdir',
+  siteDescription: 'My super site',
+  pathPrefix: '/sub/-/dir',
   plugins: []
 }