+++ gridsome/__tests__/app.spec.js
@@ -71,44 +71,44 @@ test('setup webpack server config', async () => {
   expect(config.resolve.alias['gridsome$']).toEqual(path.resolve(__dirname, '../app/index.js'))
 })
 
-// test('setup style loader options', async () => {
-//   const app = await createApp(context, {
-//     localConfig: {
-//       css: {
-//         loaderOptions: {
-//           css: { url: false },
-//           sass: { data: '@import "variables.sass";' },
-//           scss: { data: '@import "variables.scss";' },
-//           less: { strictMath: true },
-//           stylus: { use: ['plugin'] },
-//           postcss: {
-//             ident: 'postcss',
-//             plugins: ['plugin']
-//           }
-//         }
-//       }
-//     }
-//   }, BOOTSTRAP_CONFIG)
+test('setup style loader options', async () => {
+  const app = await createApp(context, {
+    localConfig: {
+      css: {
+        loaderOptions: {
+          css: { url: false },
+          sass: { data: '@import "variables.sass";' },
+          scss: { data: '@import "variables.scss";' },
+          less: { strictMath: true },
+          stylus: { use: ['plugin'] },
+          postcss: {
+            ident: 'postcss',
+            plugins: ['plugin']
+          }
+        }
+      }
+    }
+  }, BOOTSTRAP_CONFIG)
 
-//   const chain = await createClientConfig(app)
-//   const oneOf = ['normal', 'modules']
+  const chain = await createClientConfig(app)
+  const oneOf = ['normal', 'modules']
 
-//   oneOf.forEach(oneOf => {
-//     const css = chain.module.rule('css').oneOf(oneOf).use('css-loader').toConfig()
-//     const postcss = chain.module.rule('postcss').oneOf(oneOf).use('postcss-loader').toConfig()
-//     const sass = chain.module.rule('sass').oneOf(oneOf).use('sass-loader').toConfig()
-//     const scss = chain.module.rule('scss').oneOf(oneOf).use('sass-loader').toConfig()
-//     const less = chain.module.rule('less').oneOf(oneOf).use('less-loader').toConfig()
-//     const stylus = chain.module.rule('stylus').oneOf(oneOf).use('stylus-loader').toConfig()
+  oneOf.forEach(oneOf => {
+    const css = chain.module.rule('css').oneOf(oneOf).use('css-loader').toConfig()
+    const postcss = chain.module.rule('postcss').oneOf(oneOf).use('postcss-loader').toConfig()
+    const sass = chain.module.rule('sass').oneOf(oneOf).use('sass-loader').toConfig()
+    const scss = chain.module.rule('scss').oneOf(oneOf).use('sass-loader').toConfig()
+    const less = chain.module.rule('less').oneOf(oneOf).use('less-loader').toConfig()
+    const stylus = chain.module.rule('stylus').oneOf(oneOf).use('stylus-loader').toConfig()
 
-//     expect(css.options.url).toEqual(false)
-//     expect(postcss.options.ident).toEqual('postcss')
-//     expect(postcss.options.plugins).toHaveLength(2)
-//     expect(postcss.options.plugins[0]).toEqual('plugin')
-//     expect(postcss.options.plugins[1]).toBeInstanceOf(Function)
-//     expect(sass.options.data).toEqual('@import "variables.sass";')
-//     expect(scss.options.data).toEqual('@import "variables.scss";')
-//     expect(less.options.strictMath).toEqual(true)
-//     expect(stylus.options.use[0]).toEqual('plugin')
-//   })
-// })
+    expect(css.options.url).toEqual(false)
+    expect(postcss.options.ident).toEqual('postcss')
+    expect(postcss.options.plugins).toHaveLength(2)
+    expect(postcss.options.plugins[0]).toEqual('plugin')
+    expect(postcss.options.plugins[1]).toBeInstanceOf(Function)
+    expect(sass.options.data).toEqual('@import "variables.sass";')
+    expect(scss.options.data).toEqual('@import "variables.scss";')
+    expect(less.options.strictMath).toEqual(true)
+    expect(stylus.options.use[0]).toEqual('plugin')
+  })
+})