+++ gridsome/app/pages/404.vue
@@ -1,3 +1,3 @@
 <template>
-  <h1>404 - not found default</h1>
+  <h1>404 - not found</h1>
 </template>