+++ gridsome/app/router.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import config from '~/.temp/config.js'
-import { routes, NotFound } from '~/.temp/routes.js'
+import routes from '~/.temp/routes.js'
 
 Vue.use(Router)
 
@@ -9,11 +9,7 @@ const router = new Router({
   base: process.env.PUBLIC_PATH,
   mode: 'history',
   fallback: false,
-  routes: [...routes, {
-    path: '*',
-    name: '404',
-    component: NotFound
-  }],
+  routes,
 
   scrollBehavior (to, from, saved) {
     if (saved) {