+++ gridsome/lib/app/createRoutes.js
@@ -5,6 +5,7 @@ const { info } = require('../utils/log')
 const {
   PAGED_ROUTE,
   STATIC_ROUTE,
+  NOT_FOUND_ROUTE,
   STATIC_TEMPLATE_ROUTE,
   DYNAMIC_TEMPLATE_ROUTE
 } = require('../utils/constants')
@@ -17,10 +18,7 @@ module.exports = ({ store, config }) => {
   const pagedPages = []
   const staticTemplates = []
   const dynamicTemplates = []
-
-  const notFoundComponent = notFoundPage
-    ? notFoundPage.component
-    : path.join(config.appPath, 'pages', '404.vue')
+  const specialPages = []
 
   pages.forEach(page => {
     const name = camelCase(page.path.replace(/\//g, ' ')) || 'home'
@@ -87,13 +85,24 @@ module.exports = ({ store, config }) => {
     }
   })
 
-  return {
-    notFoundComponent,
-    pages: [
-      ...staticPages,
-      ...pagedPages,
-      ...staticTemplates,
-      ...dynamicTemplates
-    ]
-  }
+  specialPages.push({
+    path: '*',
+    name: '404',
+    directoryIndex: false,
+    type: NOT_FOUND_ROUTE,
+    component: notFoundPage
+      ? notFoundPage.component
+      : path.join(config.appPath, 'pages', '404.vue'),
+    pageQuery: notFoundPage
+      ? notFoundPage.pageQuery
+      : {}
+  })
+
+  return [
+    ...staticPages,
+    ...pagedPages,
+    ...staticTemplates,
+    ...dynamicTemplates,
+    ...specialPages
+  ]
 }